+++ src/Service/Handler/SentryExtraInformationHandler.php
@@ -6,14 +6,16 @@ namespace Paysera\LoggingExtraBundle\Service\Handler;
 
 use Monolog\Handler\HandlerWrapper;
 use Monolog\Handler\ProcessableHandlerTrait;
+use Monolog\LogRecord;
 use Sentry\State\Scope;
+
 use function Sentry\withScope;
 
 final class SentryExtraInformationHandler extends HandlerWrapper
 {
     use ProcessableHandlerTrait;
 
-    public function handle(array $record): bool
+    public function handle(LogRecord $record): bool
     {
         if (!$this->isHandling($record)) {
             return false;
@@ -25,15 +27,15 @@ final class SentryExtraInformationHandler extends HandlerWrapper
         $record['formatted'] = $this->getFormatter()->format($record);
 
         withScope(function (Scope $scope) use ($record, &$result): void {
-            if (isset($record['context']['extra']) && \is_array($record['context']['extra'])) {
-                foreach ($record['context']['extra'] as $key => $value) {
+            if (isset($record['context']) && \is_array($record['context'])) {
+                foreach ($record['context'] as $key => $value) {
                     $scope->setExtra((string) $key, $value);
                 }
             }
 
-            if (isset($record['context']['tags']) && \is_array($record['context']['tags'])) {
-                foreach ($record['context']['tags'] as $key => $value) {
-                    $scope->setTag($key, $value);
+            if (isset($record['extra']) && \is_array($record['extra'])) {
+                foreach ($record['extra'] as $key => $value) {
+                    $scope->setTag($key, (string)$value);
                 }
             }